+++ packages/tailwindcss-language-server/src/projects.ts
@@ -444,16 +444,26 @@ export async function createProjectService(
     let applyComplexClasses: any
 
     try {
-      let tailwindcssPath = await resolver.resolveJsId('tailwindcss', configDir)
-      let tailwindcssPkgPath = await resolver.resolveJsId('tailwindcss/package.json', configDir)
+      let tailwindcssPkgPath = await resolver.resolveCjsId('tailwindcss/package.json', configDir)
       let tailwindDir = path.dirname(tailwindcssPkgPath)
       tailwindcssVersion = require(tailwindcssPkgPath).version
 
-      let features = supportedFeatures(tailwindcssVersion)
-      log(`supported features: ${JSON.stringify(features)}`)
+      // Loading via `await import(…)` with the Yarn PnP API is not possible
+      if (await resolver.hasPnP()) {
+        let tailwindcssPath = await resolver.resolveCjsId('tailwindcss', configDir)
 
-      tailwindcssPath = pathToFileURL(tailwindcssPath).href
-      tailwindcss = await import(tailwindcssPath)
+        tailwindcss = require(tailwindcssPath)
+      } else {
+        let tailwindcssPath = await resolver.resolveJsId('tailwindcss', configDir)
+        let tailwindcssURL = pathToFileURL(tailwindcssPath).href
+
+        tailwindcss = await import(tailwindcssURL)
+      }
+
+      // TODO: The module should be loaded in the project locator
+      // and this should be determined there and passed in instead
+      let features = supportedFeatures(tailwindcssVersion, tailwindcss)
+      log(`supported features: ${JSON.stringify(features)}`)
 
       if (!features.includes('css-at-theme')) {
         tailwindcss = tailwindcss.default ?? tailwindcss
@@ -484,15 +494,18 @@ export async function createProjectService(
         log('CSS-based configuration is not supported before Tailwind CSS v4')
         state.enabled = false
         enabled = false
-        // CSS-based configuration is not supported before Tailwind CSS v4 so bail
-        // TODO: Fall back to built-in version of v4
+
+        // The fallback to a bundled v4 is in the catch block
         return
       }
 
-      const postcssPath = resolveFrom(tailwindDir, 'postcss')
-      const postcssPkgPath = resolveFrom(tailwindDir, 'postcss/package.json')
+      const postcssPath = await resolver.resolveCjsId('postcss', tailwindDir)
+      const postcssPkgPath = await resolver.resolveCjsId('postcss/package.json', tailwindDir)
       const postcssDir = path.dirname(postcssPkgPath)
-      const postcssSelectorParserPath = resolveFrom(tailwindDir, 'postcss-selector-parser')
+      const postcssSelectorParserPath = await resolver.resolveCjsId(
+        'postcss-selector-parser',
+        tailwindDir,
+      )
 
       postcssVersion = require(postcssPkgPath).version
 
@@ -665,6 +678,31 @@ export async function createProjectService(
         } catch (_) {}
       }
     } catch (error) {
+      if (projectConfig.config.source === 'css') {
+        // @ts-ignore
+        let tailwindcss = await import('tailwindcss-v4')
+        let tailwindcssVersion = require('tailwindcss-v4/package.json').version
+        let features = supportedFeatures(tailwindcssVersion, tailwindcss)
+
+        log('Failed to load workspace modules.')
+        log(`Using bundled version of \`tailwindcss\`: v${tailwindcssVersion}`)
+
+        state.configPath = configPath
+        state.version = tailwindcssVersion
+        state.isCssConfig = true
+        state.v4 = true
+        state.v4Fallback = true
+        state.jit = true
+        state.modules = {
+          tailwindcss: { version: tailwindcssVersion, module: tailwindcss },
+          postcss: { version: null, module: null },
+          resolveConfig: { module: null },
+          loadConfig: { module: null },
+        }
+
+        return tryRebuild()
+      }
+
       let util = await import('node:util')
 
       console.error(util.format(error))
@@ -781,6 +819,7 @@ export async function createProjectService(
           state.modules.tailwindcss.module,
           state.configPath,
           css,
+          state.v4Fallback ?? false,
         )
 
         state.designSystem = designSystem
@@ -1058,6 +1097,7 @@ export async function createProjectService(
           state.modules.tailwindcss.module,
           state.configPath,
           css,
+          state.v4Fallback ?? false,
         )
       } catch (err) {
         console.error(err)