+++ packages/tailwindcss-language-service/src/util/docsUrl.ts
@@ -9,6 +9,14 @@ export function docsUrl(version: string, paths: string | string[]): string {
   }
   if (semver.gte(version, '1.99.0')) {
     major = 2
+    url = 'https://v2.tailwindcss.com/docs/'
+  }
+  if (semver.gte(version, '2.99.0')) {
+    major = 3
+    url = 'https://v3.tailwindcss.com/docs/'
+  }
+  if (semver.gte(version, '3.99.0')) {
+    major = 4
     url = 'https://tailwindcss.com/docs/'
   }
   const path = Array.isArray(paths) ? paths[major] || paths[paths.length - 1] : paths