+++ packages/tailwindcss-language-service/src/util/state.ts
@@ -128,6 +128,7 @@ export interface State {
   }
 
   v4?: boolean
+  v4Fallback?: boolean
   designSystem?: DesignSystem
 
   browserslist?: string[]